# Conflicts:
#	tcfcli/cmds/deploy/cli.py
#	tcfcli/common/daily_task.py

+++ tcfcli/common/daily_task.py
@@ -46,13 +46,16 @@ def daily_task():
 
             now_time = time.strftime("%Y-%m-%d")  # day
             version_time = uc.version_time
-
+            allow_report = uc.section_map[UserConfig.OTHERS]['allow_report']
+            #print allow_report
             if now_time != version_time:
                 url = "https://service-qgphxt7y-1253970226.gz.apigw.tencentcs.com/release/client_daily_task"
-                statistics = StatisticsConfigure()
-                statistics.read_data()
-                post_data = statistics.get_data()
-                statistics.delete_data()
+                post_data = None
+                if allow_report == 'True':
+                    statistics = StatisticsConfigure()
+                    statistics.read_data()
+                    post_data = statistics.get_data()
+                    statistics.delete_data()
                 # print post_data
                 if not post_data:
                     post_data = {}
